+++ source/extensions/omni.flux.utils.widget/omni/flux/utils/widget/tree_widget/widget.py
@@ -18,6 +18,7 @@
 from __future__ import annotations
 
 import abc
+import math
 from typing import TYPE_CHECKING, Callable
 
 from omni import ui
@@ -25,6 +26,9 @@ from omni.flux.utils.common import Event as _Event
 from omni.flux.utils.common import EventSubscription as _EventSubscription
 from omni.flux.utils.common import reset_default_attrs as _reset_default_attrs
 from omni.flux.utils.common.decorators import limit_recursion as _limit_recursion
+
+from .delegate import AlternatingRowDelegate as _AlternatingRowDelegate
+from .model import AlternatingRowModel as _AlternatingRowModel
 
 if TYPE_CHECKING:
     from .delegate import TreeDelegateBase as _TreeDelegateBase
@@ -138,6 +142,137 @@ class TreeWidget(ui.TreeView):
                     yield child
                     if recursive:
                         yield from self.iter_visible_children(items=[child], recursive=recursive)
+
+    def destroy(self):
+        _reset_default_attrs(self)
+
+
+class AlternatingRowWidget:
+    def __init__(
+        self,
+        header_height: int,
+        row_height: int,
+        scrollbar_spacing: bool = True,
+        model: _AlternatingRowModel | None = None,
+        delegate: _AlternatingRowDelegate | None = None,
+    ):
+        """
+        A tree widget to be layered below another widget to create an alternating row effect.
+
+        Notes:
+            When using this widget, make sure to:
+
+            - Sync the frame height to make sure the entire frame is filled with alternating row backgrounds.
+                The `computed_content_size_changed_fn` callback of the frame wrapping the foreground tree can be used
+
+            - Sync the scroll position whenever the foreground tree's scroll position changes.
+                The `scroll_y_changed_fn` callback of the wrapping scrolling frame can be used to sync the scroll
+                position
+
+            - Refresh the widget whenever the number of items in the foreground tree changes.
+                The `subscribe_item_changed_fn` callback of the Tree Model can be used to refresh the widget
+
+        Example:
+            >>> def _on_content_size_changed(self):
+            >>>     self._alternating_row_widget.sync_frame_height(self._tree_widget.computed_height)
+            >>>
+            >>> def _on_item_changed(self):
+            >>>     # Expect self._tree_model.item_count to return the total number of items (recursively) in the model
+            >>>     self._alternating_row_widget.refresh(self._tree_model.item_count)
+            >>>
+            >>> def _build_ui(self):
+            >>>     with ui.ZStack():
+            >>>         # Build the AlternatingRowWidget
+            >>>         self._alternating_row_widget = AlternatingRowWidget(header_height, row_height)
+            >>>         # Build the Foreground TreeView wrapped in a ScrollingFrame
+            >>>         tree_scroll_frame = ui.ScrollingFrame(
+            >>>             computed_content_size_changed_fn=self._on_content_size_changed,
+            >>>             scroll_y_changed_fn=self._alternating_row_widget.sync_scrolling_frame,
+            >>>         )
+            >>>         with tree_scroll_frame:
+            >>>             self._tree_model = TreeModel(...)
+            >>>             self._tree_delegate = TreeDelegate(...)
+            >>>             self._item_changed_sub = self._tree_model.subscribe_item_changed_fn(self._on_item_changed)
+            >>>             self._tree_widget = ui.TreeView(
+            >>>                 self.tree.model,
+            >>>                 delegate=self.tree.delegate,
+            >>>             )
+        """
+
+        self._default_attr = {
+            "_header_height": None,
+            "_row_height": None,
+            "_model": None,
+            "_delegate": None,
+            "_scroll_frame": None,
+            "_tree": None,
+            "_row_count": None,
+            "_min_row_count": None,
+        }
+        for attr, value in self._default_attr.items():
+            setattr(self, attr, value)
+
+        self._header_height = header_height
+        self._row_height = row_height
+        self._model = model or _AlternatingRowModel()
+        self._delegate = delegate or _AlternatingRowDelegate(row_height, scrollbar_spacing=scrollbar_spacing)
+
+        self._scroll_frame = None
+        self._tree = None
+
+        self._row_count = 0
+        self._min_row_count = 0
+
+        self._build_ui()
+
+    def _build_ui(self):
+        """
+        Build the alternating row TreeView UI elements
+        """
+        self._scroll_frame = ui.ScrollingFrame(
+            name="TreePanelBackground",
+            vertical_scrollbar_policy=ui.ScrollBarPolicy.SCROLLBAR_ALWAYS_OFF,
+            horizontal_scrollbar_policy=ui.ScrollBarPolicy.SCROLLBAR_ALWAYS_OFF,
+        )
+        with self._scroll_frame:
+            with ui.VStack():
+                ui.Spacer(height=ui.Pixel(self._header_height))
+                self._tree = ui.TreeView(
+                    self._model,
+                    delegate=self._delegate,
+                    header_visible=False,
+                )
+
+    def sync_frame_height(self, frame_height: int):
+        """
+        Sync the frame height to make sure the entire frame is filled with alternating row backgrounds.
+
+        Args:
+            frame_height: The height of the frame this widget should fill.
+        """
+        self._min_row_count = math.ceil(frame_height / self._row_height)
+        # Expand the tree is the new minimum row count is bigger than the current row count
+        if self._row_count < self._min_row_count:
+            self.refresh()
+
+    def sync_scrolling_frame(self, position: float):
+        """
+        Sync the scroll position whenever the foreground tree's scroll position changes.
+
+        Args:
+            position: The new scroll position
+        """
+        self._scroll_frame.scroll_y = position
+
+    def refresh(self, item_count: int = 0):
+        """
+        Refresh the widget whenever the number of items in the foreground tree changes.
+
+        Args:
+            item_count: The number of items in the foreground tree
+        """
+        self._row_count = max(item_count, self._min_row_count)
+        self._model.refresh(self._row_count)
 
     def destroy(self):
         _reset_default_attrs(self)